Created Util Method to Upload Images To Cloudinary
